+++ LLama/LLama/Util/HTTPUtil/LLAHttpUtil.m
@@ -82,7 +82,7 @@ static NSString *const httpBaseURL = @"https://api.hillama.com";
     if (progressBlock) {
         
         [manager setDownloadTaskDidWriteDataBlock:^(NSURLSession * _Nonnull session, NSURLSessionDownloadTask * _Nonnull downloadTask, int64_t bytesWritten, int64_t totalBytesWritten, int64_t totalBytesExpectedToWrite) {
-            progressBlock(bytesWritten,totalBytesWritten,totalBytesExpectedToWrite);
+            progressBlock((NSUInteger)bytesWritten,totalBytesWritten,totalBytesExpectedToWrite);
         }];
     }
